Account for the NMVContext cdf as well
